% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_utils.R
\name{find_in_fasta}
\alias{find_in_fasta}
\title{R wrapper to ssearch}
\usage{
find_in_fasta(sequences, target_fasta, ssearch_exe, seq_names = NULL,
  min_eval = 0.01)
}
\arguments{
\item{sequences}{sequences to be searched}

\item{target_fasta}{path to fasta file containing the target sequences}

\item{ssearch_exe}{location of the ssearch executable (or alias in your path)}

\item{seq_names}{optional names to be given to the sequences}

\item{min_eval}{minimum E value to be reported, defaults to 0.01}
}
\value{
tabular console output of ssearch as a character vector (element per line)
}
\description{
Generates a temporary fasta file with the sequences provided and returns the
matches in the provided fasta file.
}
\details{
It wraps sseach to do the search and then returns the console output
}
\examples{
\donttest{find_in_fasta(c('PEPTIDE'), 'arabidopsis.fasta', '~/usr/bin/ssearch36')}
}
