% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jspp_utils.R
\name{make_exp_design}
\alias{make_exp_design}
\title{Title}
\usage{
make_exp_design(eset, factor_vector = c(WT = "WT", KO = "KO"),
  reference_factor = "KO", batch_vector = NULL)
}
\arguments{
\item{factor_vector}{}
}
\description{
Title
}
