% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_utils.R
\name{make_tmp_fasta}
\alias{make_tmp_fasta}
\title{Make a temporary fasta file}
\usage{
make_tmp_fasta(sequences, seq_names = NULL)
}
\arguments{
\item{sequences}{Character vector of sequences to be written}

\item{seq_names}{Character vector of names to be used in the fasta header (defaults to the sequences)}
}
\value{
Path to the temporary fasta
}
\description{
Make a temporary fasta file
}
\examples{
make_tmp_fasta(c("PEPTIDE", "TIDALES"))
make_tmp_fasta(c("PEPTIDE", "TIDALES"), seq_names = c("Pep1", "Pep2"))
}
