% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_utils.R
\name{parse_ssearch_out}
\alias{parse_ssearch_out}
\title{Parses the console output of ssearch}
\usage{
parse_ssearch_out(ssearch_out)
}
\arguments{
\item{ssearch_out}{tabular console output of ssearch as a character vector (element per line)}
}
\value{
data.frame with output of the ssearch search
}
\description{
It is designed to parse que tabular out of ssearch (flag -m 8C)
}
\examples{
\donttest{
    ss_out <- find_in_fasta(c('PEPTIDE'), 'arabidopsis.fasta', '~/usr/bin/ssearch36')
    parse_ssearch_out(ss_out)
    }
}
